requirements.txt
discord.py
yt-dlp
pillow
